dealing with the type limits and math
