+++ packages/user/DiffAdjust/service/src/tests/mod.rs
@@ -1,0 +1,7 @@
+#![allow(non_snake_case)]
+
+mod adjustment;
+mod limits;
+mod windows;
+
+pub(crate) const ONE_MILLION: f64 = 1_000_000.0;
